+++ src/nice_int/nice_u32.rs
@@ -63,7 +63,6 @@ macro_rules! inner {
 pub type NiceU32 = NiceWrapper<SIZE>;
 
 super::nice_default!(NiceU32, inner!(b','), SIZE);
-super::nice_from!(NiceU32, u32);
 super::nice_from_nz!(NiceU32, NonZeroU32);
 super::nice_parse!(NiceU32, u32);